refactor User Interactive Flow Handling Functions
